requirements.txt
Flask
Flask_Cors
openai==1.0.0
requests
python-dotenv
